requirement.txt
Django==2.2.5
djangorestframework==3.13.1
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
